Initial Commit
The updated net.py includes more comments, added some more options to set for the parameters of the neural networks, calculation of tau inside instead of using a static value, changed what the predict function returns to make it possible to get the mean and standard deviation for each prediction and a small tweak to make it work with the latest  Keras and Tensorflow. Including an example I created to test the changes (very different from Yarin's experiments and not meant to compare to them).